+++ app/src/main/java/nodomain/freeyourgadget/gadgetbridge/devices/huami/Huami2021Coordinator.java
@@ -202,6 +202,7 @@ public abstract class Huami2021Coordinator extends HuamiCoordinator {
         settings.add(R.xml.devicesettings_sleep_mode);
         settings.add(R.xml.devicesettings_liftwrist_display_sensitivity_with_smart);
         settings.add(R.xml.devicesettings_password);
+        settings.add(R.xml.devicesettings_huami2021_watchface);
         settings.add(R.xml.devicesettings_always_on_display);
         settings.add(R.xml.devicesettings_screen_timeout);
         if (supportsAutoBrightness(device)) {